Rename translation types file
