+++ src/common/layout/languageDropdown/__tests__/LanguageDropdown.test.tsx
@@ -3,7 +3,7 @@ import { act } from 'react-dom/test-utils';
 import { Dropdown } from 'reactstrap';
 
 import i18n from '../../../../locales/i18n';
-import { LocaleOpts } from '../../../types/intl';
+import { LocaleOpts } from '../../../types/translation';
 import LanguageDropdown from '../LanguageDropdown';
 
 describe('LanguageDropdown', () => {
