--- src/common/types/intl.ts
@@ -1,7 +1,0 @@
-export enum LocaleOpts {
-  FI = 'fi',
-  EN = 'en',
-  SV = 'sv',
-}
-
-export type FormatMessage = (msg: { id: string }) => string;
