+++ src/common/types/translation.ts
@@ -1,0 +1,8 @@
+export enum LocaleOpts {
+  FI = 'fi',
+  EN = 'en',
+  SV = 'sv',
+}
+
+// Used to differ translations for the same key
+export type TContext = 'berth' | 'winter';
